src/test/java/dev/ryandunaway/didemo/controllers/PropertyInjectedControllerTest.java
package dev.ryandunaway.didemo.controllers;

import dev.ryandunaway.didemo.services.GreetingServiceImpl;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

public class PropertyInjectedControllerTest {

    private PropertyInjectedController propertyInjectedController;

    @Before
    public void setUp() throws Exception {
        this.propertyInjectedController = new PropertyInjectedController();
        this.propertyInjectedController.grettingService = new GreetingServiceImpl();
    }

    @Test
    public void sayHello() {
        Assert.assertEquals(GreetingServiceImpl.HELLO_GURUS, propertyInjectedController.sayHello());
    }
}